[lib] Use ‘size_t’ for ‘svz_array_t’ dimensions and indices.
U ≡ "Convert type to ‘size_t’ for local var".
* src/libserveez/array.c (svz_array_t): Convert
type to ‘size_t’ for members ‘.size’, ‘.capacity’.
(svz_array_cretae): Convert type of arg ‘capacity’ to ‘size_t’.
(svz_array_destroy): Likewise for local var.
(svz_array_ensure_capacity): Likewise for arg ‘size’.
(svz_array_get, svz_array_set)
(svz_array_del): Likewise for arg ‘index’.
(svz_array_size): Return ‘size_t’.
(svz_array_strdup): U.
* src/libserveez/array.h (svz_array_create)
(svz_array_get, svz_array_set, svz_array_del)
(svz_array_size): Update func decls.
(svz_array_foreach): Remove cast to ‘unsigned long’.
* src/libserveez/binding.c (svz_binding_find_server)
(svz_server_portcfgs, svz_binding_find, svz_binding_contains)
(svz_binding_join, svz_server_bind, svz_binding_contains_server)
(svz_sock_servers, svz_binding_filter_pipe, svz_binding_filter_net)
(svz_pp_server_bindings): U.
* src/libserveez/codec.c (svz_foreach_codec, svz_codec_get)
(svz_codec_register, svz_codec_unregister, svz_codec_sock_detect): U.
* src/libserveez/coserver/coserver.c (svz_coserver_send_request): U.
(svz_foreach_coserver, svz_coserver_activate, svz_coserver_count)
(svz_coserver_delete, svz_coserver_disconnect, svz_coserver_close_pipes)
(svz_coserver_destroy, svz_coserver_check, svz_coserver_create): U.
* src/libserveez/core.c (svz_file_del, svz_file_closeall): U.
* src/libserveez/dynload.c (dyn_libraries): Change type to ‘size_t’.
(svz_dynload_path_set, svz_dynload_path_get, dyn_load_library)
(dyn_unload_library, dyn_load_symbol): U.
* src/libserveez/icmp-socket.c (svz_icmp_check_request): U.
* src/libserveez/interface.c (svz_foreach_interface)
[__MINGW32__] (svz_interface_collect)
(svz_interface_add, svz_interface_get)
(svz_interface_search, svz_interface_check): U.
* src/libserveez/server-core.c (svz_sock_check_access): U.
* src/libserveez/server-socket.c (svz_sock_check_frequency): U.
* src/libserveez/server.c (svz_foreach_servertype, svz_servertype_add): U.
(svz_sock_del_server): U; return ‘size_t’.
(svz_servertype_get, svz_servertype_finalize): U.
* src/libserveez/socket.c (svz_sock_detect_proto): U.
* src/libserveez/udp-socket.c (svz_udp_check_request): U.

U ≡ "Convert type to ‘size_t’ for local var".
* src/ctrl-server/control-proto.c (ctrl_stat_id): U.
* src/foo-server/foo-proto.c (foo_connect_socket, foo_info_server): U.
* src/guile-api.c (guile_server_listeners, guile_server_clients): U.
* src/guile-server.c
(guile_servertype_config_print): U; remove cast to ‘int’.
* src/guile.c (guile_access_interfaces)
(guile_strarray_to_guile, guile_intarray_to_guile): U.
* src/irc-server/irc-config.c (irc_parse_config_lines): U.
* src/irc-server/irc-proto.c (irc_check_tcp_bindings): U.
* src/irc-server/irc-server.c (irc_connect_servers): U.
* src/nut-server/gnutella.h (nut_config_t): Convert type to
‘size_t’ for members ‘.search_index’.
* src/nut-server/gnutella.c (nut_init, nut_server_notify): U.
(nut_idle_searching): Remove cast to ‘unsigned long’.
(nut_info_server): Likewise; also, U.
* src/nut-server/nut-transfer.c
(nut_init_transfer): U; remove cast to ‘unsigned long’.
* src/prog-server/prog-server.h (prog_config_t):
Change type to ‘size_t’ of member ‘.frequency’.
* src/prog-server/prog-server.c
(prog_check_frequency): U; remove cast to ‘unsigned long’.
(prog_passthrough, prog_init): U.

U ≡ "Convert type to ‘size_t’ for local var".
* test/btdt.c (array_size_correct): Delete func.
(array_popcount): U; return ‘size_t’.
(array_idx): U.
(array_main): U; compare rv of ‘svz_array_size’ directly.
(hash_main): U.

+++ src/guile-api.c
@@ -657,7 +657,7 @@ guile_server_listeners (SCM server)
   svz_server_t *xserver = NULL;
   svz_array_t *listeners;
   char *str;
-  unsigned long i;
+  size_t i;
   SCM list = SCM_EOL;
 
   /* Server instance name given?  */
@@ -697,7 +697,7 @@ guile_server_clients (SCM server)
   svz_array_t *clients;
   svz_socket_t *sock;
   char *str;
-  unsigned long i;
+  size_t i;
   SCM list = SCM_EOL;
 
   /* If the server instance name is given, try to translate it.  */
